Move ILabStatus token to tokens.ts
